requirements.txt
torch>=1.0
torchvision
pillow
flask